// fe-reviewer/src/app/model/company.model.ts
export class Company {
  // idCompany : string;
  nameCompany: string;
  addrCompany: string;
  webCompany: string;
  telCompany: string;
  dtCreated: Date;
  idAccount: string;
  imgAvatarCompany: string;
  imgPanelCompany: string;
  emailCompany: string;
  password: string;
  confirmPassword: string;
}